refactor(scale): define scale in specFont.js, not in designSpecs.js
The scale variable is about font-size beyond 728px wide screens, not about layout per se.
