+++ src/blocks/Main.js
@@ -1,14 +1,9 @@
 import styled from 'styled-components';
 
 import {mediaQuery, twoColumns} from 'src/utils/breakpoints';
-import {
-  flexbox,
-  formNumberLarge,
-  header,
-  page,
-  scale,
-} from 'src/utils/designSpecs';
+import {flexbox, formNumberLarge, header, page} from 'src/utils/designSpecs';
 import remify from 'src/utils/remify';
+import {scale} from 'src/utils/specFont';
 
 const Main = styled.main`
   padding-top: ${remify(header.height + header.borderBottomWidth)};
