+++ src/blocks/WrapperForms.js
@@ -2,9 +2,9 @@ import styled from 'styled-components';
 
 import {mediaQuery} from 'src/utils/breakpoints';
 import {boxSize} from 'src/utils/layout';
-import {page, rgbHslWrapper, scale} from 'src/utils/designSpecs';
+import {page, rgbHslWrapper} from 'src/utils/designSpecs';
 import remify from 'src/utils/remify';
-
+import {scale} from 'src/utils/specFont';
 const WrapperForms = styled.section``;
 
 WrapperForms.MarginBetweenColorCodeAndHexRgbHsl = styled.div`
