+++ src/components/LinkHelpIcon.js
@@ -1,10 +1,11 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import styled from 'styled-components';
-import {capheight, scale} from 'src/utils/designSpecs';
+
 import {color} from 'src/utils/color';
 import {mediaQuery} from 'src/utils/breakpoints';
 import remify from 'src/utils/remify';
+import {scale} from 'src/utils/specFont';
 
 // color scheme
 const black = color.input.background;
